pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.93.0",
    "psycopg2-binary>=2.9.10",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.41",
    "google-genai>=1.24.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "sift-stack-py>=0.7.0",
    "scikit-learn>=1.7.0",
    "matplotlib>=3.10.3",
    "seaborn>=0.13.2",
    "numpy>=2.3.1",
]
